split Hypervolume Algorithms and clean up. Tests need to be updated. #54
